update pose estimation
