--- pose_estimation/ap-tool.sh
@@ -1,27 +1,0 @@
-#!/bin/bash
-
-
-start_ap() {
-	sudo rfkill unblock wlan
-
-	sudo ifconfig wlan0 192.168.1.1/24 up
-	sudo hostapd /etc/hostapd/hostapd.conf -B
-
-	sudo systemctl stop systemd-resolved.service
-	sudo dnsmasq -C /etc/dnsmasq.conf
-}
-
-stop_ap() {
-	sudo kill -9 `pidof hostapd`
-	sudo kill -9 `pidof dnsmasq`
-	sudo ifconfig wlan0 down
-}
-
-case $1 in
-	start)
-		start_ap
-		;;
-	stop)
-		stop_ap
-		;;
-esac
