+++ pose_estimation/regions.json
@@ -1,1 +1,1 @@
-[[[130, 172], [196, 163], [201, 265], [130, 262]], [[265, 116], [366, 114], [385, 260], [272, 274]], [[414, 146], [466, 148], [471, 194], [419, 194]], [[103, 135], [91, 188], [50, 196], [57, 145]]]
+[[[324, 433], [524, 319], [653, 408], [567, 476], [359, 478]], [[294, 311], [368, 285], [398, 325], [321, 363]], [[271, 252], [332, 231], [360, 250], [296, 276]], [[212, 218], [289, 203], [303, 220], [222, 246]]]
