+++ pose_estimation/start.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python3 /home/nvidia/people/peopletcp.py > /dev/null
